int size

+++ src/apps/testapps/testCompact.c
@@ -159,7 +159,7 @@ SUITE(compact) {
         // Arbitrary index
         setH3Index(&h3, res, 0, 2);
 
-        int arrSize = H3_EXPORT(cellToChildrenSize)(h3, res + 1) + 1;
+        int64_t arrSize = H3_EXPORT(cellToChildrenSize)(h3, res + 1) + 1;
         H3Index* children = calloc(arrSize, sizeof(H3Index));
 
         H3_EXPORT(h3ToChildren)(h3, res + 1, children);
@@ -183,7 +183,7 @@ SUITE(compact) {
         // Arbitrary pentagon parent cell
         setH3Index(&h3, res, 4, 0);
 
-        int arrSize = H3_EXPORT(cellToChildrenSize)(h3, res + 1) + 1;
+        int64_t arrSize = H3_EXPORT(cellToChildrenSize)(h3, res + 1) + 1;
         H3Index* children = calloc(arrSize, sizeof(H3Index));
 
         H3_EXPORT(h3ToChildren)(h3, res + 1, children);
@@ -209,7 +209,7 @@ SUITE(compact) {
         // Arbitrary index
         setH3Index(&h3, res, 0, 2);
 
-        int arrSize = H3_EXPORT(cellToChildrenSize)(h3, res + 1);
+        int64_t arrSize = H3_EXPORT(cellToChildrenSize)(h3, res + 1);
         H3Index* children = calloc(arrSize, sizeof(H3Index));
 
         H3_EXPORT(h3ToChildren)(h3, res + 1, children);
